Rename cmd_util to env_util
