+++ docs/modules/ppo.rst
@@ -55,7 +55,7 @@ Train a PPO agent on ``Pendulum-v0`` using 4 environments.
 
   from stable_baselines3 import PPO
   from stable_baselines3.ppo import MlpPolicy
-  from stable_baselines3.common.cmd_util import make_vec_env
+  from stable_baselines3.common.env_util import make_vec_env
 
   # Parallel environments
   env = make_vec_env('CartPole-v1', n_envs=4)
